feat: 회의 생성 페이지 구조 변경
